+++ zcash_client_sqlite/src/wallet/init/migrations/orchard_shardtree.rs
@@ -4,7 +4,7 @@
 use std::collections::HashSet;
 
 use rusqlite::{named_params, OptionalExtension};
-use schemer_rusqlite::RusqliteMigration;
+use schemerz_rusqlite::RusqliteMigration;
 use tracing::debug;
 use uuid::Uuid;
 use zcash_client_backend::data_api::scanning::ScanPriority;
@@ -21,7 +21,7 @@ pub(super) struct Migration<P> {
     pub(super) params: P,
 }
 
-impl<P> schemer::Migration for Migration<P> {
+impl<P> schemerz::Migration for Migration<P> {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }