+++ zcash_client_sqlite/CHANGELOG.md
@@ -11,6 +11,9 @@ and this library adheres to Rust's notion of
 - Exposed `AccountId::from_u32` and `AccountId::as_u32` conversions under the
   `unstable` feature flag.
 
+### Changed
+- Migrated from `schemer` to our fork `schemerz`.
+
 ## [0.12.1] - 2024-10-10
 
 ### Fixed